+++ store/rem/domain/memory.go
@@ -7,8 +7,8 @@ import (
 	"errors"
 	"strconv"
 
-	"github.com/mrsirg97-rgb/looper/store/lazy"
-	"github.com/mrsirg97-rgb/looper/store/sqlx"
+	"github.com/mrsirg97-rgb/rig/store/lazy"
+	"github.com/mrsirg97-rgb/rig/store/sqlx"
 )
 
 type Memory struct {